Minor refactoring of QuickwitConfig.
