+++ quickwit/quickwit-config/src/quickwit_config/serialize.rs
@@ -17,15 +17,12 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-use std::collections::{HashMap, HashSet};
-use std::env;
+use std::collections::HashMap;
 use std::net::{IpAddr, SocketAddr};
-use std::path::PathBuf;
 use std::str::FromStr;
 
 use anyhow::{bail, Context};
-use byte_unit::Byte;
-use quickwit_common::net::{find_private_ip, Host, HostAddr};
+use quickwit_common::net::{find_private_ip, Host};
 use quickwit_common::new_coolid;
 use quickwit_common::uri::Uri;
 use serde::{Deserialize, Serialize};
@@ -35,9 +32,7 @@ use crate::config_value::ConfigValue;
 use crate::qw_env_vars::*;
 use crate::service::QuickwitService;
 use crate::templating::render_config;
-use crate::{validate_identifier, ConfigFormat};
-
-pub const DEFAULT_QW_CONFIG_PATH: &str = "config/quickwit.yaml";
+use crate::{validate_identifier, ConfigFormat, IndexerConfig, QuickwitConfig, SearcherConfig};
 
 pub const DEFAULT_CLUSTER_ID: &str = "quickwit-default-cluster";
 
@@ -50,6 +45,22 @@ fn default_cluster_id() -> ConfigValue<String, QW_CLUSTER_ID> {
 
 fn default_node_id() -> ConfigValue<String, QW_NODE_ID> {
     ConfigValue::with_default(new_coolid("node"))
+}
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize, PartialEq)]
+struct List(Vec<String>);
+
+impl FromStr for List {
+    type Err = anyhow::Error;
+
+    fn from_str(list_str: &str) -> Result<Self, Self::Err> {
+        let list = list_str
+            .split(',')
+            .map(|elem| elem.trim().to_string())
+            .filter(|elem| !elem.is_empty())
+            .collect();
+        Ok(List(list))
+    }
 }
 
 fn default_enabled_services() -> ConfigValue<List, QW_ENABLED_SERVICES> {
@@ -101,131 +112,37 @@ fn default_index_root_uri(data_dir_uri: &Uri) -> Uri {
     data_dir_uri.join("indexes").expect("Failed to create default index root URI. This should never happen! Please, report on https://github.com/quickwit-oss/quickwit/issues.")
 }
 
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct IndexerConfig {
-    #[doc(hidden)]
-    #[serde(default = "IndexerConfig::default_enable_opentelemetry_otlp_service")]
-    pub enable_opentelemetry_otlp_service: bool,
-    #[serde(default = "IndexerConfig::default_split_store_max_num_bytes")]
-    pub split_store_max_num_bytes: Byte,
-    #[serde(default = "IndexerConfig::default_split_store_max_num_splits")]
-    pub split_store_max_num_splits: usize,
-    #[serde(default = "IndexerConfig::default_max_concurrent_split_uploads")]
-    pub max_concurrent_split_uploads: usize,
+pub async fn load_quickwit_config_with_env(
+    config_format: ConfigFormat,
+    config_content: &[u8],
+    env_vars: &HashMap<String, String>,
+) -> anyhow::Result<QuickwitConfig> {
+    let rendered_config_content = render_config(config_content)?;
+    let versioned_quickwit_config: VersionedQuickwitConfig =
+        config_format.parse(rendered_config_content.as_bytes())?;
+    let quickwit_config_builder: QuickwitConfigBuilder = versioned_quickwit_config.into();
+    let config = quickwit_config_builder.build_and_validate(env_vars).await?;
+    Ok(config)
 }
 
-impl IndexerConfig {
-    fn default_enable_opentelemetry_otlp_service() -> bool {
-        false
-    }
-
-    fn default_max_concurrent_split_uploads() -> usize {
-        12
-    }
-
-    pub fn default_split_store_max_num_bytes() -> Byte {
-        Byte::from_bytes(100_000_000_000) // 100G
-    }
-
-    pub fn default_split_store_max_num_splits() -> usize {
-        1_000
-    }
-
-    #[cfg(any(test, feature = "testsuite"))]
-    pub fn for_test() -> anyhow::Result<Self> {
-        let indexer_config = IndexerConfig {
-            enable_opentelemetry_otlp_service: true,
-            split_store_max_num_bytes: Byte::from_bytes(1_000_000),
-            split_store_max_num_splits: 3,
-            max_concurrent_split_uploads: 4,
-        };
-        Ok(indexer_config)
-    }
+#[derive(Debug, Deserialize)]
+#[serde(tag = "version")]
+enum VersionedQuickwitConfig {
+    #[serde(rename = "3")]
+    V3(QuickwitConfigBuilder),
 }
 
-impl Default for IndexerConfig {
-    fn default() -> Self {
-        Self {
-            enable_opentelemetry_otlp_service: Self::default_enable_opentelemetry_otlp_service(),
-            split_store_max_num_bytes: Self::default_split_store_max_num_bytes(),
-            split_store_max_num_splits: Self::default_split_store_max_num_splits(),
-            max_concurrent_split_uploads: Self::default_max_concurrent_split_uploads(),
+impl From<VersionedQuickwitConfig> for QuickwitConfigBuilder {
+    fn from(versioned_quickwit_config: VersionedQuickwitConfig) -> Self {
+        match versioned_quickwit_config {
+            VersionedQuickwitConfig::V3(quickwit_config_builder) => quickwit_config_builder,
         }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct SearcherConfig {
-    #[doc(hidden)]
-    #[serde(default = "SearcherConfig::default_enable_jaeger_service")]
-    pub enable_jaeger_service: bool,
-    #[serde(default = "SearcherConfig::default_fast_field_cache_capacity")]
-    pub fast_field_cache_capacity: Byte,
-    #[serde(default = "SearcherConfig::default_split_footer_cache_capacity")]
-    pub split_footer_cache_capacity: Byte,
-    #[serde(default = "SearcherConfig::default_max_num_concurrent_split_searches")]
-    pub max_num_concurrent_split_searches: usize,
-    #[serde(default = "SearcherConfig::default_max_num_concurrent_split_streams")]
-    pub max_num_concurrent_split_streams: usize,
-}
-
-impl SearcherConfig {
-    fn default_enable_jaeger_service() -> bool {
-        false
-    }
-
-    fn default_fast_field_cache_capacity() -> Byte {
-        Byte::from_bytes(1_000_000_000) // 1G
-    }
-
-    fn default_split_footer_cache_capacity() -> Byte {
-        Byte::from_bytes(500_000_000) // 500M
-    }
-
-    fn default_max_num_concurrent_split_searches() -> usize {
-        100
-    }
-
-    fn default_max_num_concurrent_split_streams() -> usize {
-        100
-    }
-}
-
-impl Default for SearcherConfig {
-    fn default() -> Self {
-        Self {
-            enable_jaeger_service: Self::default_enable_jaeger_service(),
-            fast_field_cache_capacity: Self::default_fast_field_cache_capacity(),
-            split_footer_cache_capacity: Self::default_split_footer_cache_capacity(),
-            max_num_concurrent_split_streams: Self::default_max_num_concurrent_split_streams(),
-            max_num_concurrent_split_searches: Self::default_max_num_concurrent_split_searches(),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Default, Serialize, Deserialize, PartialEq)]
-struct List(Vec<String>);
-
-impl FromStr for List {
-    type Err = anyhow::Error;
-
-    fn from_str(list_str: &str) -> Result<Self, Self::Err> {
-        let list = list_str
-            .split(',')
-            .map(|elem| elem.trim().to_string())
-            .filter(|elem| !elem.is_empty())
-            .collect();
-        Ok(List(list))
     }
 }
 
 #[derive(Debug, Deserialize, PartialEq)]
 #[serde(deny_unknown_fields)]
 struct QuickwitConfigBuilder {
-    #[serde(default)]
-    version: ConfigValue<String, QW_NONE>,
     #[serde(default = "default_cluster_id")]
     cluster_id: ConfigValue<String, QW_CLUSTER_ID>,
     #[serde(default = "default_node_id")]
@@ -255,13 +172,10 @@ struct QuickwitConfigBuilder {
 }
 
 impl QuickwitConfigBuilder {
-    fn from_uri(uri: &Uri, config_content: &[u8]) -> anyhow::Result<Self> {
-        let config_format = ConfigFormat::sniff_from_uri(uri)?;
-        let rendered_config = render_config(uri, config_content)?;
-        config_format.parse(rendered_config.as_bytes())
-    }
-
-    pub async fn build(self, env_vars: &HashMap<String, String>) -> anyhow::Result<QuickwitConfig> {
+    pub async fn build_and_validate(
+        self,
+        env_vars: &HashMap<String, String>,
+    ) -> anyhow::Result<QuickwitConfig> {
         let enabled_services = self
             .enabled_services
             .resolve(env_vars)?
@@ -319,8 +233,7 @@ impl QuickwitConfigBuilder {
             .resolve_optional(env_vars)?
             .unwrap_or_else(|| default_index_root_uri(&data_dir_uri));
 
-        Ok(QuickwitConfig {
-            version: self.version.resolve(env_vars)?,
+        let quickwit_config = QuickwitConfig {
             cluster_id: self.cluster_id.resolve(env_vars)?,
             node_id: self.node_id.resolve(env_vars)?,
             enabled_services,
@@ -335,178 +248,114 @@ impl QuickwitConfigBuilder {
             default_index_root_uri,
             indexer_config: self.indexer_config,
             searcher_config: self.searcher_config,
-        })
+        };
+
+        validate(&quickwit_config)?;
+        Ok(quickwit_config)
     }
 }
 
-#[derive(Clone, Debug, Serialize)]
-pub struct QuickwitConfig {
-    pub version: String,
-    pub cluster_id: String,
-    pub node_id: String,
-    pub enabled_services: HashSet<QuickwitService>,
-    pub rest_listen_addr: SocketAddr,
-    pub gossip_listen_addr: SocketAddr,
-    pub grpc_listen_addr: SocketAddr,
-    pub gossip_advertise_addr: SocketAddr,
-    pub grpc_advertise_addr: SocketAddr,
-    pub peer_seeds: Vec<String>,
-    pub data_dir_path: PathBuf,
-    pub metastore_uri: Uri,
-    pub default_index_root_uri: Uri,
-    pub indexer_config: IndexerConfig,
-    pub searcher_config: SearcherConfig,
+fn validate(quickwit_config: &QuickwitConfig) -> anyhow::Result<()> {
+    validate_identifier("Cluster ID", &quickwit_config.cluster_id)?;
+    validate_identifier("Node ID", &quickwit_config.node_id)?;
+
+    if quickwit_config.cluster_id == DEFAULT_CLUSTER_ID {
+        warn!(
+            cluster_id=%DEFAULT_CLUSTER_ID,
+            "Cluster ID is not set, falling back to default value."
+        );
+    }
+    if quickwit_config.peer_seeds.is_empty() {
+        warn!("Peer seed list is empty.");
+    }
+    if !quickwit_config.data_dir_path.exists() {
+        bail!(
+            "Data dir `{}` does not exist.",
+            quickwit_config.data_dir_path.display()
+        );
+    }
+    Ok(())
 }
 
-impl QuickwitConfig {
-    /// Parses and validates a [`QuickwitConfig`] from a given URI and config content.
-    pub async fn load(config_uri: &Uri, config_content: &[u8]) -> anyhow::Result<Self> {
-        let env_vars = env::vars().collect::<HashMap<_, _>>();
-        let config = Self::load_with_env(config_uri, config_content, &env_vars).await?;
-        Ok(config)
-    }
-
-    pub async fn load_with_env(
-        config_uri: &Uri,
-        config_content: &[u8],
-        env_vars: &HashMap<String, String>,
-    ) -> anyhow::Result<Self> {
-        let config_builder = QuickwitConfigBuilder::from_uri(config_uri, config_content)?;
-        let config = config_builder.build(env_vars).await?;
-        config.validate()?;
-        Ok(config)
-    }
-
-    pub fn validate(&self) -> anyhow::Result<()> {
-        validate_identifier("Cluster ID", &self.cluster_id)?;
-        validate_identifier("Node ID", &self.node_id)?;
-
-        if self.cluster_id == DEFAULT_CLUSTER_ID {
-            warn!(
-                cluster_id=%DEFAULT_CLUSTER_ID,
-                "Cluster ID is not set, falling back to default value."
-            );
-        }
-        if self.peer_seeds.is_empty() {
-            warn!("Peer seed list is empty.");
-        }
-        if !self.data_dir_path.exists() {
-            bail!(
-                "Data dir `{}` does not exist.",
-                self.data_dir_path.display()
-            );
-        }
-        Ok(())
-    }
-
-    /// Returns the list of peer seed addresses. The addresses MUST NOT be resolved. Otherwise, the
-    /// DNS-based discovery mechanism implemented in Chitchat will not work correctly.
-    pub async fn peer_seed_addrs(&self) -> anyhow::Result<Vec<String>> {
-        let mut peer_seed_addrs = Vec::new();
-        let default_gossip_port = self.gossip_listen_addr.port();
-
-        // We want to pass non-resolved addresses to Chitchat but still want to resolve them for
-        // validation purposes. Additionally, we need to append a default port if necessary and
-        // finally return the addresses as strings, which is tricky for IPv6. We let the logic baked
-        // in `HostAddr` handle this complexity.
-        for peer_seed in &self.peer_seeds {
-            let peer_seed_addr = HostAddr::parse_with_default_port(peer_seed, default_gossip_port)?;
-            if let Err(error) = peer_seed_addr.resolve().await {
-                warn!(peer_seed = %peer_seed_addr, error = ?error, "Failed to resolve peer seed address.");
-                continue;
-            }
-            peer_seed_addrs.push(peer_seed_addr.to_string())
-        }
-        if !self.peer_seeds.is_empty() && peer_seed_addrs.is_empty() {
-            bail!(
-                "Failed to resolve any of the peer seed addresses: `{}`",
-                self.peer_seeds.join(", ")
-            )
-        }
-        Ok(peer_seed_addrs)
-    }
-
-    #[cfg(any(test, feature = "testsuite"))]
-    pub fn for_test() -> Self {
-        let enabled_services = QuickwitService::supported_services();
-
-        let listen_address = Host::default();
-        let rest_listen_port = quickwit_common::net::find_available_tcp_port()
-            .expect("The OS should almost always find an available port.");
-        let rest_listen_addr = listen_address
-            .with_port(rest_listen_port)
-            .to_socket_addr()
-            .expect("The default host should be an IP address.");
-        let gossip_listen_addr = listen_address
-            .with_port(rest_listen_port)
-            .to_socket_addr()
-            .expect("The default host should be an IP address.");
-        let grpc_listen_port = quickwit_common::net::find_available_tcp_port()
-            .expect("The OS should almost always find an available port.");
-        let grpc_listen_addr = listen_address
-            .with_port(grpc_listen_port)
-            .to_socket_addr()
-            .expect("The default host should be an IP address.");
-
-        let data_dir_uri = default_data_dir_uri().unwrap();
-        let data_dir_path = data_dir_uri
-            .filepath()
-            .expect("The default data dir should be valid directory path.")
-            .to_path_buf();
-        let metastore_uri = default_metastore_uri(&data_dir_uri);
-        let default_index_root_uri = default_index_root_uri(&data_dir_uri);
-
+impl Default for QuickwitConfigBuilder {
+    fn default() -> Self {
         Self {
-            version: "3".to_string(),
-            cluster_id: default_cluster_id().unwrap(),
-            node_id: default_node_id().unwrap(),
-            enabled_services,
-            gossip_advertise_addr: gossip_listen_addr,
-            grpc_advertise_addr: grpc_listen_addr,
-            rest_listen_addr,
-            gossip_listen_addr,
-            grpc_listen_addr,
-            peer_seeds: Vec::new(),
-            data_dir_path,
-            metastore_uri,
-            default_index_root_uri,
+            cluster_id: default_cluster_id(),
+            node_id: default_node_id(),
+            enabled_services: default_enabled_services(),
+            listen_address: default_listen_address(),
+            rest_listen_port: default_rest_listen_port(),
+            gossip_listen_port: ConfigValue::none(),
+            grpc_listen_port: ConfigValue::none(),
+            advertise_address: ConfigValue::none(),
+            peer_seeds: ConfigValue::with_default(List::default()),
+            data_dir_uri: default_data_dir_uri(),
+            metastore_uri: ConfigValue::none(),
+            default_index_root_uri: ConfigValue::none(),
             indexer_config: IndexerConfig::default(),
             searcher_config: SearcherConfig::default(),
         }
     }
 }
 
+#[cfg(any(test, feature = "testsuite"))]
+pub fn quickwit_config_for_test() -> QuickwitConfig {
+    let enabled_services = QuickwitService::supported_services();
+
+    let listen_address = Host::default();
+    let rest_listen_port = quickwit_common::net::find_available_tcp_port()
+        .expect("The OS should almost always find an available port.");
+    let rest_listen_addr = listen_address
+        .with_port(rest_listen_port)
+        .to_socket_addr()
+        .expect("The default host should be an IP address.");
+    let gossip_listen_addr = listen_address
+        .with_port(rest_listen_port)
+        .to_socket_addr()
+        .expect("The default host should be an IP address.");
+    let grpc_listen_port = quickwit_common::net::find_available_tcp_port()
+        .expect("The OS should almost always find an available port.");
+    let grpc_listen_addr = listen_address
+        .with_port(grpc_listen_port)
+        .to_socket_addr()
+        .expect("The default host should be an IP address.");
+
+    let data_dir_uri = default_data_dir_uri().unwrap();
+    let data_dir_path = data_dir_uri
+        .filepath()
+        .expect("The default data dir should be valid directory path.")
+        .to_path_buf();
+    let metastore_uri = default_metastore_uri(&data_dir_uri);
+    let default_index_root_uri = default_index_root_uri(&data_dir_uri);
+
+    QuickwitConfig {
+        cluster_id: default_cluster_id().unwrap(),
+        node_id: default_node_id().unwrap(),
+        enabled_services,
+        gossip_advertise_addr: gossip_listen_addr,
+        grpc_advertise_addr: grpc_listen_addr,
+        rest_listen_addr,
+        gossip_listen_addr,
+        grpc_listen_addr,
+        peer_seeds: Vec::new(),
+        data_dir_path,
+        metastore_uri,
+        default_index_root_uri,
+        indexer_config: IndexerConfig::default(),
+        searcher_config: SearcherConfig::default(),
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use std::env;
-    use std::net::{IpAddr, Ipv4Addr, SocketAddr};
+    use std::net::Ipv4Addr;
     use std::path::Path;
 
+    use byte_unit::Byte;
     use itertools::Itertools;
 
     use super::*;
-
-    impl Default for QuickwitConfigBuilder {
-        fn default() -> Self {
-            Self {
-                version: ConfigValue::none(),
-                cluster_id: default_cluster_id(),
-                node_id: default_node_id(),
-                enabled_services: default_enabled_services(),
-                listen_address: default_listen_address(),
-                rest_listen_port: default_rest_listen_port(),
-                gossip_listen_port: ConfigValue::none(),
-                grpc_listen_port: ConfigValue::none(),
-                advertise_address: ConfigValue::none(),
-                peer_seeds: ConfigValue::with_default(List::default()),
-                data_dir_uri: default_data_dir_uri(),
-                metastore_uri: ConfigValue::none(),
-                default_index_root_uri: ConfigValue::none(),
-                indexer_config: IndexerConfig::default(),
-                searcher_config: SearcherConfig::default(),
-            }
-        }
-    }
 
     fn get_config_filepath(config_filename: &str) -> String {
         format!(
@@ -520,14 +369,11 @@ mod tests {
     async fn test_quickwit_config_parse_aux(config_format: ConfigFormat) -> anyhow::Result<()> {
         let config_filepath =
             get_config_filepath(&format!("quickwit.{config_format:?}").to_lowercase());
-        let config_uri = Uri::from_str(&config_filepath)?;
         let file = std::fs::read_to_string(&config_filepath).unwrap();
-        let config = QuickwitConfigBuilder::from_uri(&config_uri, file.as_bytes())?
-            .build(&HashMap::new())
-            .await?;
-        assert_eq!(config.version, "3");
+        let env_vars = HashMap::default();
+        let config =
+            load_quickwit_config_with_env(config_format, file.as_bytes(), &env_vars).await?;
         assert_eq!(config.cluster_id, "quickwit-cluster");
-
         assert_eq!(config.enabled_services.len(), 2);
 
         assert!(config.enabled_services.contains(&QuickwitService::Janitor));
@@ -614,10 +460,14 @@ mod tests {
     #[tokio::test]
     async fn test_config_contains_wrong_values() {
         let config_filepath = get_config_filepath("quickwit.wrongkey.yaml");
-        let config_uri = Uri::from_str(&config_filepath).unwrap();
         let config_str = std::fs::read_to_string(&config_filepath).unwrap();
-        let parsing_error =
-            QuickwitConfigBuilder::from_uri(&config_uri, config_str.as_bytes()).unwrap_err();
+        let parsing_error = super::load_quickwit_config_with_env(
+            ConfigFormat::Yaml,
+            config_str.as_bytes(),
+            &Default::default(),
+        )
+        .await
+        .unwrap_err();
         assert!(format!("{parsing_error:?}")
             .contains("unknown field `max_num_concurrent_split_searchs`"));
     }
@@ -637,9 +487,13 @@ mod tests {
     #[tokio::test]
     async fn test_quickwit_config_default_values_minimal() {
         let config_yaml = "version: 3";
-        let config_builder = serde_yaml::from_str::<QuickwitConfigBuilder>(config_yaml).unwrap();
-        let config = config_builder.build(&HashMap::new()).await.unwrap();
-        assert_eq!(config.version, "3");
+        let config = load_quickwit_config_with_env(
+            ConfigFormat::Yaml,
+            config_yaml.as_bytes(),
+            &Default::default(),
+        )
+        .await
+        .unwrap();
         assert_eq!(config.cluster_id, DEFAULT_CLUSTER_ID);
         assert!(config.node_id.starts_with("node-"));
         assert_eq!(
@@ -681,7 +535,6 @@ mod tests {
     #[tokio::test]
     async fn test_quickwit_config_env_var_override() {
         let config_yaml = "version: 3";
-        let config_builder = serde_yaml::from_str::<QuickwitConfigBuilder>(config_yaml).unwrap();
         let mut env_vars = HashMap::new();
         env_vars.insert("QW_CLUSTER_ID".to_string(), "test-cluster".to_string());
         env_vars.insert("QW_NODE_ID".to_string(), "test-node".to_string());
@@ -707,7 +560,10 @@ mod tests {
             "QW_DEFAULT_INDEX_ROOT_URI".to_string(),
             "s3://quickwit-indexes/prod".to_string(),
         );
-        let config = config_builder.build(&env_vars).await.unwrap();
+        let config =
+            load_quickwit_config_with_env(ConfigFormat::Yaml, config_yaml.as_bytes(), &env_vars)
+                .await
+                .unwrap();
         assert_eq!(config.cluster_id, "test-cluster");
         assert_eq!(config.node_id, "test-node");
         assert_eq!(config.enabled_services.len(), 2);
@@ -761,14 +617,18 @@ mod tests {
     async fn test_quickwwit_config_default_values_storage() {
         let config_yaml = r#"
             version: 3
-            node_id: 1
+            node_id: "node1"
             metastore_uri: postgres://username:password@host:port/db
         "#;
-        let config_builder = serde_yaml::from_str::<QuickwitConfigBuilder>(config_yaml).unwrap();
-        let config = config_builder.build(&HashMap::new()).await.unwrap();
-        assert_eq!(config.version, "3");
+        let config = load_quickwit_config_with_env(
+            ConfigFormat::Yaml,
+            config_yaml.as_bytes(),
+            &Default::default(),
+        )
+        .await
+        .unwrap();
         assert_eq!(config.cluster_id, DEFAULT_CLUSTER_ID);
-        assert_eq!(config.node_id, "1");
+        assert_eq!(config.node_id, "node1");
         assert_eq!(
             config.metastore_uri,
             "postgres://username:password@host:port/db"
@@ -782,9 +642,13 @@ mod tests {
             metastore_uri: postgres://username:password@host:port/db
             data_dir: /opt/quickwit/data
         "#;
-        let config_builder = serde_yaml::from_str::<QuickwitConfigBuilder>(config_yaml).unwrap();
-        let config = config_builder.build(&HashMap::new()).await.unwrap();
-        assert_eq!(config.version, "3");
+        let config = load_quickwit_config_with_env(
+            ConfigFormat::Yaml,
+            config_yaml.as_bytes(),
+            &Default::default(),
+        )
+        .await
+        .unwrap();
         assert_eq!(
             config.metastore_uri,
             "postgres://username:password@host:port/db"
@@ -796,7 +660,6 @@ mod tests {
     #[tokio::test]
     async fn test_quickwit_config_validate() {
         let config_filepath = get_config_filepath("quickwit.toml");
-        let config_uri = Uri::from_str(&config_filepath).unwrap();
         let file_content = std::fs::read_to_string(&config_filepath).unwrap();
 
         let data_dir_path = env::current_dir().unwrap();
@@ -805,40 +668,40 @@ mod tests {
             "QW_DATA_DIR".to_string(),
             data_dir_path.to_string_lossy().to_string(),
         );
-        let config = QuickwitConfig::load_with_env(&config_uri, file_content.as_bytes(), &env_vars)
-            .await
-            .unwrap();
-        assert!(config.validate().is_ok());
+        assert!(load_quickwit_config_with_env(
+            ConfigFormat::Toml,
+            file_content.as_bytes(),
+            &env_vars,
+        )
+        .await
+        .is_ok());
     }
 
     #[tokio::test]
     async fn test_peer_socket_addrs() {
         {
             let quickwit_config = QuickwitConfigBuilder {
-                version: ConfigValue::for_test("3".to_string()),
                 rest_listen_port: ConfigValue::for_test(1789),
                 ..Default::default()
             }
-            .build(&HashMap::new())
+            .build_and_validate(&HashMap::new())
             .await
             .unwrap();
             assert!(quickwit_config.peer_seed_addrs().await.unwrap().is_empty());
         }
         {
             let quickwit_config = QuickwitConfigBuilder {
-                version: ConfigValue::for_test("3".to_string()),
                 rest_listen_port: ConfigValue::for_test(1789),
                 peer_seeds: ConfigValue::for_test(List(vec!["unresolvable-host".to_string()])),
                 ..Default::default()
             }
-            .build(&HashMap::new())
+            .build_and_validate(&HashMap::new())
             .await
             .unwrap();
             assert!(quickwit_config.peer_seed_addrs().await.is_err());
         }
         {
             let quickwit_config = QuickwitConfigBuilder {
-                version: ConfigValue::for_test("3".to_string()),
                 rest_listen_port: ConfigValue::for_test(1789),
                 peer_seeds: ConfigValue::for_test(List(vec![
                     "unresolvable-host".to_string(),
@@ -849,7 +712,7 @@ mod tests {
                 ])),
                 ..Default::default()
             }
-            .build(&HashMap::new())
+            .build_and_validate(&HashMap::new())
             .await
             .unwrap();
             assert_eq!(
@@ -868,11 +731,10 @@ mod tests {
     async fn test_socket_addr_ports() {
         {
             let quickwit_config = QuickwitConfigBuilder {
-                version: ConfigValue::for_test("3".to_string()),
                 listen_address: default_listen_address(),
                 ..Default::default()
             }
-            .build(&HashMap::new())
+            .build_and_validate(&HashMap::new())
             .await
             .unwrap();
             assert_eq!(
@@ -890,12 +752,11 @@ mod tests {
         }
         {
             let quickwit_config = QuickwitConfigBuilder {
-                version: ConfigValue::for_test("3".to_string()),
                 listen_address: default_listen_address(),
                 rest_listen_port: ConfigValue::for_test(1789),
                 ..Default::default()
             }
-            .build(&HashMap::new())
+            .build_and_validate(&HashMap::new())
             .await
             .unwrap();
             assert_eq!(
@@ -913,14 +774,13 @@ mod tests {
         }
         {
             let quickwit_config = QuickwitConfigBuilder {
-                version: ConfigValue::for_test("3".to_string()),
                 listen_address: default_listen_address(),
                 rest_listen_port: ConfigValue::for_test(1789),
                 gossip_listen_port: ConfigValue::for_test(1889),
                 grpc_listen_port: ConfigValue::for_test(1989),
                 ..Default::default()
             }
-            .build(&HashMap::new())
+            .build_and_validate(&HashMap::new())
             .await
             .unwrap();
             assert_eq!(
@@ -957,7 +817,13 @@ mod tests {
             node_id: 1
             metastore_uri: ''
         "#;
-            serde_yaml::from_str::<QuickwitConfigBuilder>(config_yaml).unwrap_err();
+            assert!(load_quickwit_config_with_env(
+                ConfigFormat::Yaml,
+                config_yaml.as_bytes(),
+                &Default::default()
+            )
+            .await
+            .is_err());
         }
         {
             let config_yaml = r#"
@@ -966,7 +832,13 @@ mod tests {
             metastore_uri: postgres://username:password@host:port/db
             default_index_root_uri: ''
         "#;
-            serde_yaml::from_str::<QuickwitConfigBuilder>(config_yaml).unwrap_err();
+            assert!(load_quickwit_config_with_env(
+                ConfigFormat::Yaml,
+                config_yaml.as_bytes(),
+                &Default::default()
+            )
+            .await
+            .is_err());
         }
     }
 
@@ -977,29 +849,41 @@ mod tests {
                 version: 3
                 data_dir: /opt/quickwit/data
             "#;
-            let config_builder =
-                serde_yaml::from_str::<QuickwitConfigBuilder>(config_yaml).unwrap();
-            let config = config_builder.build(&HashMap::new()).await.unwrap();
-            assert_eq!(config.data_dir_path, PathBuf::from("/opt/quickwit/data"));
+            let config = load_quickwit_config_with_env(
+                ConfigFormat::Yaml,
+                config_yaml.as_bytes(),
+                &HashMap::default(),
+            )
+            .await
+            .unwrap();
+            assert_eq!(&config.data_dir_path, Path::new("/opt/quickwit/data"));
         }
         {
             let config_yaml = r#"
                 version: 3
                 data_dir: file:///opt/quickwit/data
             "#;
-            let config_builder =
-                serde_yaml::from_str::<QuickwitConfigBuilder>(config_yaml).unwrap();
-            let config = config_builder.build(&HashMap::new()).await.unwrap();
-            assert_eq!(config.data_dir_path, PathBuf::from("/opt/quickwit/data"));
+            let config = load_quickwit_config_with_env(
+                ConfigFormat::Yaml,
+                config_yaml.as_bytes(),
+                &HashMap::default(),
+            )
+            .await
+            .unwrap();
+            assert_eq!(&config.data_dir_path, Path::new("/opt/quickwit/data"));
         }
         {
             let config_yaml = r#"
                 version: 3
                 data_dir: s3://indexes/foo
             "#;
-            let config_builder =
-                serde_yaml::from_str::<QuickwitConfigBuilder>(config_yaml).unwrap();
-            let error = config_builder.build(&HashMap::new()).await.unwrap_err();
+            let error = load_quickwit_config_with_env(
+                ConfigFormat::Yaml,
+                config_yaml.as_bytes(),
+                &HashMap::default(),
+            )
+            .await
+            .unwrap_err();
             assert!(error.to_string().contains("Data dir must be located"));
         }
     }
