+++ quickwit/quickwit-config/src/templating.rs
@@ -22,7 +22,6 @@ use std::collections::HashMap;
 use anyhow::{bail, Context, Result};
 use new_string_template::template::Template;
 use once_cell::sync::Lazy;
-use quickwit_common::uri::Uri;
 use regex::Regex;
 use tracing::debug;
 
@@ -33,10 +32,9 @@ static TEMPLATE_ENV_VAR_CAPTURE: Lazy<Regex> = Lazy::new(|| {
     Regex::new(r"\$\{\s*([A-Za-z0-9_]+)(?:(?::\-)([\S]+))?\s*}").expect("Failed to compile regular expression. This should never happen! Please, report on https://github.com/quickwit-oss/quickwit/issues.")
 });
 
-pub fn render_config(config_uri: &Uri, config_content: &[u8]) -> Result<String> {
-    let template_str = String::from_utf8(config_content.to_vec()).with_context(|| {
-        format!("Config file `{config_uri}` contains invalid UTF-8 characters.")
-    })?;
+pub fn render_config(config_content: &[u8]) -> Result<String> {
+    let template_str = std::str::from_utf8(config_content)
+        .context("Config file contains invalid UTF-8 characters.")?;
 
     let mut values = HashMap::new();
 
@@ -70,10 +68,10 @@ pub fn render_config(config_uri: &Uri, config_content: &[u8]) -> Result<String> 
         };
         values.insert(env_var_key, substitution_value);
     }
-    let template = Template::new(&template_str).with_regex(&TEMPLATE_ENV_VAR_CAPTURE);
+    let template = Template::new(template_str).with_regex(&TEMPLATE_ENV_VAR_CAPTURE);
     let rendered = template
         .render(&values)
-        .with_context(|| format!("Failed to render templated config file `{config_uri}`."))?;
+        .context("Failed to render templated config file")?;
     Ok(rendered)
 }
 
@@ -81,13 +79,7 @@ pub fn render_config(config_uri: &Uri, config_content: &[u8]) -> Result<String> 
 mod test {
     use std::env;
 
-    use once_cell::sync::Lazy;
-    use quickwit_common::uri::Uri;
-
     use super::render_config;
-
-    static TEST_URI: Lazy<Uri> =
-        Lazy::new(|| Uri::from_well_formed("file://config.yaml".to_string()));
 
     #[test]
     fn test_template_render() {
@@ -96,7 +88,7 @@ mod test {
             "TEST_TEMPLATE_RENDER_ENV_VAR_PLEASE_DONT_NOTICE",
             "s3://test-bucket/metastore",
         );
-        let rendered = render_config(&TEST_URI, config_content).unwrap();
+        let rendered = render_config(config_content).unwrap();
         std::env::remove_var("TEST_TEMPLATE_RENDER_ENV_VAR_PLEASE_DONT_NOTICE");
         assert_eq!(rendered, "metastore_uri: s3://test-bucket/metastore");
     }
@@ -109,22 +101,22 @@ mod test {
         );
         {
             let config_content = b"metastore_uri: ${TEST_TEMPLATE_RENDER_WHITESPACE_QW_TEST}";
-            let rendered = render_config(&TEST_URI, config_content).unwrap();
+            let rendered = render_config(config_content).unwrap();
             assert_eq!(rendered, "metastore_uri: s3://test-bucket/metastore");
         }
         {
             let config_content = b"metastore_uri: ${TEST_TEMPLATE_RENDER_WHITESPACE_QW_TEST  }";
-            let rendered = render_config(&TEST_URI, config_content).unwrap();
+            let rendered = render_config(config_content).unwrap();
             assert_eq!(rendered, "metastore_uri: s3://test-bucket/metastore");
         }
         {
             let config_content = b"metastore_uri: ${   TEST_TEMPLATE_RENDER_WHITESPACE_QW_TEST}";
-            let rendered = render_config(&TEST_URI, config_content).unwrap();
+            let rendered = render_config(config_content).unwrap();
             assert_eq!(rendered, "metastore_uri: s3://test-bucket/metastore");
         }
         {
             let config_content = b"metastore_uri: ${  TEST_TEMPLATE_RENDER_WHITESPACE_QW_TEST    }";
-            let rendered = render_config(&TEST_URI, config_content).unwrap();
+            let rendered = render_config(config_content).unwrap();
             assert_eq!(rendered, "metastore_uri: s3://test-bucket/metastore");
         }
     }
@@ -133,14 +125,14 @@ mod test {
     fn test_template_render_default_value() {
         let config_content =
             b"metastore_uri: ${QW_NO_ENV_WITH_THIS_NAME:-s3://test-bucket/metastore}";
-        let rendered = render_config(&TEST_URI, config_content).unwrap();
+        let rendered = render_config(config_content).unwrap();
         assert_eq!(rendered, "metastore_uri: s3://test-bucket/metastore");
     }
 
     #[test]
     fn test_template_render_should_panic() {
         let config_content = b"metastore_uri: ${QW_NO_ENV_WITH_THIS_NAME}";
-        render_config(&TEST_URI, config_content).unwrap_err();
+        render_config(config_content).unwrap_err();
     }
 
     #[test]
@@ -151,7 +143,7 @@ mod test {
             "TEST_TEMPLATE_RENDER_ENV_VAR_DEFAULT_USE_ENV",
             "s3://test-bucket/metastore",
         );
-        let rendered = render_config(&TEST_URI, config_content).unwrap();
+        let rendered = render_config(config_content).unwrap();
         std::env::remove_var("TEST_TEMPLATE_RENDER_ENV_VAR_DEFAULT_USE_ENV");
         assert_eq!(rendered, "metastore_uri: s3://test-bucket/metastore");
     }
